+++ packages/react-magma-dom/src/components/TreeView/TreeItem.tsx
@@ -733,6 +733,7 @@ const TreeItemComponent = React.forwardRef<HTMLLIElement, TreeItemProps>(
         <div style={treeItemStyles}>
           <StyledTreeItem
             {...rest}
+            aria-disabled={isDisabled || null}
             aria-expanded={hasOwnTreeItems ? expanded : null}
             aria-selected={selectedItem}
             aria-checked={shouldShowCheckbox ? ariaCheckedValue : null}
@@ -873,6 +874,15 @@ const TreeItemComponent = React.forwardRef<HTMLLIElement, TreeItemProps>(
                   </Transition>
                 );
               }
+            )}
+            {isMacOS && (
+              <VisuallyHidden>
+                <Announce>
+                  {expanded
+                    ? i18n.expansionState.expanded
+                    : i18n.expansionState.collapsed}
+                </Announce>
+              </VisuallyHidden>
             )}
           </StyledTreeItem>
           {isMacOS && (